Improve as_string(), and fix test_compile_expr()
Make as_string() print a function and iterator signature,
so we know an object is wrapped in an iterator, or that
a string belongs to the name of a function.

Also, fix test_compile_expr() tests which broke after
we optimized out OP_FINISH_STRING into not emitting single strings.

+++ src/language/char_buffer.c
@@ -13,7 +13,7 @@ CharBuffer create_char_buffer() {
 }
 
 /** Appends a string of length with strncat. */
-void buffer_append_string_len(CharBuffer *buffer, const char *string, const int length) {
+void buffer_append_string_len(CharBuffer *buffer, const char *string, const size_t length) {
     if (buffer->length + length + 1 > buffer->capacity) {
         buffer->capacity += length * 2;
         buffer->text = ARRAY_REALLOC(buffer->text, buffer->capacity, sizeof(char));
